feat: Add course association to StudentsImport
The StudentsImport class now accepts a course_id parameter in its constructor. This allows associating imported students with a specific course. If a student is not already associated with the course, the import process will attach the course to the student.

This change enhances the functionality of the StudentsImport class and provides more flexibility when importing student data.